+++ src/public/virtual-keyboard-types.ts
@@ -4,7 +4,8 @@ import type {
   AlphabeticKeyboardLayout,
   VirtualKeyboardLayer,
   VirtualKeyboardLayout,
-  ActionToolbarOptions,
+  EditToolbarOptions,
+  VirtualKeyboardKeycap,
 } from './virtual-keyboard';
 
 export interface MathfieldProxy {
@@ -89,7 +90,11 @@ export type VirtualKeyboardMessage =
       alphabeticLayout?: AlphabeticKeyboardLayout;
       layers: Record<string, string | Partial<VirtualKeyboardLayer>>;
       layouts: (string | VirtualKeyboardLayout)[];
-      actionToolbar?: ActionToolbarOptions;
+      editToolbar?: EditToolbarOptions;
+      actionKeycap: string | Partial<VirtualKeyboardKeycap>;
+      shiftKeycap: string | Partial<VirtualKeyboardKeycap>;
+      backspaceKeycap: string | Partial<VirtualKeyboardKeycap>;
+      tabKeycap: string | Partial<VirtualKeyboardKeycap>;
     }
   | {
       // From proxy to VK
@@ -98,7 +103,11 @@ export type VirtualKeyboardMessage =
       alphabeticLayout?: AlphabeticKeyboardLayout;
       layers: Record<string, string | Partial<VirtualKeyboardLayer>>;
       layouts: (string | VirtualKeyboardLayout)[];
-      actionToolbar?: ActionToolbarOptions;
+      editToolbar?: EditToolbarOptions;
+      actionKeycap: string | Partial<VirtualKeyboardKeycap>;
+      shiftKeycap: string | Partial<VirtualKeyboardKeycap>;
+      backspaceKeycap: string | Partial<VirtualKeyboardKeycap>;
+      tabKeycap: string | Partial<VirtualKeyboardKeycap>;
     }
   | {
       // From proxy to VK